Moved parameters from config.yml to parameters.yml
time_steps, time_step and dielectric_par are now stored with other parameters. They can still be set in config.yml for backwards compatability and to give the user a simple way of changing them.

Moreover, config.lmp_parameters is not passed to lmp_controler anymore but a dedicated dictionary lmp_settings.